Enhance AppDomainContext and readability of tests
