+++ test/GitHub.UI.UnitTests/GitHub.UI.UnitTests.csproj
@@ -55,9 +55,9 @@
   <ItemGroup>
     <Compile Include="Helpers\AppDomainContext.cs" />
     <Compile Include="Helpers\ResourceDictionaryUtilities.cs" />
-    <Compile Include="Helpers\SharedDictionaryManagerBaseTests.cs" />
     <Compile Include="Helpers\SharedDictionaryManagerTests.cs" />
-    <Compile Include="Helpers\LoadingResourceDictionaryTests.cs" />
+    <Compile Include="Helpers\SharedDictionaryManagerIntegrationTests.cs" />
+    <Compile Include="Helpers\LoadingResourceDictionaryIntegrationTests.cs" />
     <Compile Include="Helpers\Urls.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
